fixing empty particle.data files
the indicies for where to initialize the particles were not being generated correctly so that no particles were being made. that has been fixed. there now seems to be a segfault occuring somewhere in the particle pusher.

I suspect a particle is being pushed beyond the boundaries of the grid and this causes the pusher to try to access an out of bounds index on the subsequent push of that particle.